unify on bool_from_env; fix 0 check
